requirements.txt
pypdf>=4.0.1
chromadb>=0.4.22,<0.5.0
sentence-transformers>=2.2.2,<3.0.0
python-dotenv>=0.19.0,<1.1.0
openai>=1.23.0

fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
python-multipart>=0.0.6,<0.1.0

numpy>=1.24.3,<2.0.0
PyPDF2==3.0.1

# httpx will be installed by openai or other http clients 